add support for using SAML attributes to create groups
The 'general-gid_mapping' setting may now be used to determine which of the
SAML attributes to map to the user's group id. The attribute may appear
more than once, in which case each mapped attribute will correspond to a
different group.

The autoprovision functionality has been extended so that, if the mapped
group id doesn't exist, it will be created. These are stored in the table
'*PREFIX*user_saml_groups', with the 'PREFIX*user_saml_group_user' table
used to store the membership of each provisioned group.

Also supported is the use of a specific mapped attribute value to designate
users of that group as NextCloud admins (not group admins). For example,
if the 'general-gid_admin' value is 'dept-admin' and a user has that group
amongst its group ids then that user will be provisioned as a NextCloud
admin user.

These changes should be treated as experimental: they have only been tested
with Shibboleth, not any other SAML implementations. Also, only tested
against v1.4.2 of this app with NextCloud v12.0.6. Oh, and no new unit
tests, yet.